mostly works
remaining issue, errors in log on no open document test.

project id does chane on document chagne, and when projectid is changed manually. if projectID is blank it will do the popup, only issue here is that the method used that initiates the popup does not receive the project id, but the method after that method will